+++ go/src/capturing_logger.go
