+++ go/src/config.go
@@ -1,7 +1,5 @@
 package src
 
-
 type Config interface {
-
-     sendEmailSummary() bool
+	sendEmailSummary() bool
 }
