+++ go/src/emailer.go
@@ -1,5 +1,5 @@
 package src
 
 type Emailer interface {
-    send(message string)
+	send(message string)
 }
