+++ go/src/pipeline_test.go
@@ -143,7 +143,7 @@ func TestProject_withou_tests_that_fails_build_with_email_notification(t *testin
 	assert.Equal(t, "Deployment failed", emailer.email)
 }
 
-func TestProject_withou_tests_that_fails_build_without_email_notification(t *testing.T) {
+func TestProject_without_tests_that_fails_build_without_email_notification(t *testing.T) {
 	logger, emailer, pipeline := newPipelineAndSpies(noEmail)
 
 	var project = Project{buildFailure, NO_TESTS}
