+++ go/src/project.go
@@ -5,17 +5,15 @@ type Project struct {
 	testStatus         TestStatus
 }
 
-
 func (p *Project) SetTestStatus(testStatus TestStatus) {
 	p.testStatus = testStatus
 }
-
 
 func (p Project) hasTests() bool {
 	return p.testStatus != NO_TESTS
 }
 
-func (p Project) runTests() string{
+func (p Project) runTests() string {
 	if p.testStatus == PASSING_TESTS {
 		return "success"
 	}
@@ -27,12 +25,12 @@ func (p Project) deploy() string {
 	}
 	return "failure"
 }
+
 type ProjectBuilder struct {
 	buildsSuccessfully bool
 	testStatus         TestStatus
-
 }
 
 func builder() ProjectBuilder {
-	return  ProjectBuilder{}
+	return ProjectBuilder{}
 }
